backend_python/requirements.txt
sqlalchemy
alembic
psycopg2-binary
pytest
pytest-cov
passlib[bcrypt]
python-jose
fastapi
uvicorn
python-decouple